requirements.txt
environs==9.5.0
ephem==4.1.4
python-telegram-bot==20.1
emoji==2.2.0
clarifai-grpc==9.2.0